Frigid: I now have enough money to play again, and I'm working on these as I level jobs.
Working through SMN and re-enabling some features. Some need to be enabled for the currently tested features to work, but are not tested because I'm a scrub and only level 38 on SMN. These have been marked 'UNTESTED' in their descriptions. Its not recommended to use them, but I'm also not your dad.